Fix to assignment + printing
