+++ checker/src/behavior/template_literal.rs
@@ -41,7 +41,8 @@ pub fn synthesize_template_literal<
 					let value = cast_as_string(cst, checking_data.settings.strict_casts).unwrap();
 					checking_data.types.new_constant_type(Constant::String(value))
 				} else {
-					todo!()
+					crate::utils::notify!("Need to cast to string...");
+					value
 				}
 			}
 		}
