+++ checker/src/types/subtyping.rs
@@ -10,145 +10,7 @@ use crate::{
 
 use super::{poly_types::SeedingContext, Constructor, StructureGenerics, Type};
 
-/// TODO add_property_restrictions via const generics
-pub struct BasicEquality {
-	pub add_property_restrictions: bool,
-	pub position: Span,
-}
-
-/// For subtyping
-pub trait SubtypeBehavior {
-	fn set_type_argument(
-		&mut self,
-		parameter: TypeId,
-		value: TypeId,
-		environment: &mut Environment,
-		types: &TypeStore,
-	) -> Result<(), NonEqualityReason>;
-
-	fn add_property_restrictions(&self) -> bool;
-
-	fn add_function_restriction(
-		&mut self,
-		environment: &mut Environment,
-		function_id: FunctionId,
-		function_type: FunctionType,
-	);
-
-	// TODO
-	// object reference type needs to meet constraint
-	// LHS is dependent + RHS argument
-}
-
-impl SubtypeBehavior for SeedingContext {
-	/// Does not check thingy
-	fn set_type_argument(
-		&mut self,
-		type_id: TypeId,
-		value: TypeId,
-		environment: &mut Environment,
-		types: &TypeStore,
-	) -> Result<(), NonEqualityReason> {
-		let restriction = self.type_arguments.get_restriction_for_id(type_id);
-
-		// Check restriction from call site type argument
-		if let Some((pos, restriction)) = restriction {
-			if let SubTypeResult::IsNotSubType(reason) =
-				type_is_subtype(restriction, value, None, self, environment, types)
-			{
-				return Err(NonEqualityReason::GenericRestrictionMismatch {
-					restriction,
-					reason: Box::new(reason),
-					pos,
-				});
-			}
-		}
-
-		self.type_arguments.set_id(type_id, value, types);
-
-		Ok(())
-	}
-
-	fn add_property_restrictions(&self) -> bool {
-		false
-	}
-
-	fn add_function_restriction(
-		&mut self,
-		_environment: &mut Environment,
-		function_id: FunctionId,
-		function_type: FunctionType,
-	) {
-		self.locally_held_functions.insert(function_id, function_type);
-	}
-}
-
-impl SubtypeBehavior for BasicEquality {
-	fn set_type_argument(
-		&mut self,
-		parameter: TypeId,
-		value: TypeId,
-		environment: &mut Environment,
-		types: &TypeStore,
-	) -> Result<(), NonEqualityReason> {
-		Ok(())
-	}
-
-	fn add_property_restrictions(&self) -> bool {
-		self.add_property_restrictions
-	}
-
-	fn add_function_restriction(
-		&mut self,
-		environment: &mut Environment,
-		function_id: FunctionId,
-		function_type: FunctionType,
-	) {
-		let result = environment
-			.deferred_function_constraints
-			.insert(function_id, (function_type, self.position.clone()));
-
-		debug_assert!(result.is_none());
-	}
-}
-
-#[derive(Debug)]
-pub enum SubTypeResult {
-	IsSubType,
-	IsNotSubType(NonEqualityReason),
-}
-
-// impl SubTypeResult {
-// 	type Error = NonEqualityReason;
-// }
-
-// TODO implement `?` on SupertypeResult
-
-// TODO maybe positions and extra information here
-// SomeLiteralMismatch
-// GenericParameterCollision
-#[derive(Debug)]
-pub enum NonEqualityReason {
-	Mismatch,
-	PropertiesInvalid {
-		errors: Vec<(TypeId, PropertyError)>,
-	},
-	// For function call-site type arguments
-	GenericRestrictionMismatch {
-		restriction: TypeId,
-		reason: Box<NonEqualityReason>,
-		pos: Span,
-	},
-	TooStrict,
-	/// TODO more information
-	MissingParameter,
-}
-
-#[derive(Debug)]
-pub enum PropertyError {
-	Missing,
-	Invalid { expected: TypeId, found: TypeId, mismatch: NonEqualityReason },
-}
+pub use super::{BasicEquality, NonEqualityReason, PropertyError, SubTypeResult, SubtypeBehavior};
 
 /// TODO tidy function
 /// TODO account for getters and setters
@@ -243,11 +105,10 @@ pub fn type_is_subtype<T: SubtypeBehavior>(
 				let right_func = types.functions.get(right_func).unwrap();
 				for (idx, lhs_param) in left_func.parameters.parameters.iter().enumerate() {
 					match right_func.parameters.get_type_constraint_at_index(idx) {
-						Some(ty) => {
-							// TODO order here
+						Some(right_param_ty) => {
 							let result = type_is_subtype(
-								ty,
 								lhs_param.ty,
+								right_param_ty,
 								ty_arguments,
 								behavior,
 								environment,
